+++ packages/backend-function/src/factory.ts
@@ -15,7 +15,7 @@ import { Construct } from 'constructs';
 import { NodejsFunction, OutputFormat } from 'aws-cdk-lib/aws-lambda-nodejs';
 import * as path from 'path';
 import { getCallerDirectory } from './get_caller_directory.js';
-import { Duration } from 'aws-cdk-lib';
+import { Duration, Stack } from 'aws-cdk-lib';
 import { Runtime } from 'aws-cdk-lib/aws-lambda';
 import { createRequire } from 'module';
 import { FunctionEnvironmentTranslator } from './function_env_translator.js';
@@ -27,6 +27,10 @@ import {
   functionOutputKey,
 } from '@aws-amplify/backend-output-schemas';
 import { FunctionEnvironmentTypeGenerator } from './function_env_type_generator.js';
+import { AttributionMetadataStorage } from '@aws-amplify/backend-output-storage';
+import { fileURLToPath } from 'url';
+
+const functionStackType = 'function-Lambda';
 
 /**
  * Entry point for defining a function in the Amplify ecosystem
@@ -278,6 +282,13 @@ class AmplifyFunction
       .map((line) => line.replace(/\/\/.*$/, '')) // strip out inline comments because the banner is going to be flattened into a single line
       .join('');
 
+    const functionEnvironmentTypeGenerator =
+      new FunctionEnvironmentTypeGenerator(id);
+
+    // esbuild runs as part of the NodejsFunction constructor, so we eagerly generate the process env shim without types so it can be included in the function bundle.
+    // This will be overwritten with the typed file at the end of synthesis
+    functionEnvironmentTypeGenerator.generateProcessEnvShim();
+
     const functionLambda = new NodejsFunction(scope, `${id}-lambda`, {
       entry: props.entry,
       timeout: Duration.seconds(props.timeoutSeconds),
@@ -294,7 +305,7 @@ class AmplifyFunction
       functionLambda,
       props.environment,
       backendSecretResolver,
-      new FunctionEnvironmentTypeGenerator(id)
+      functionEnvironmentTypeGenerator
     );
 
     this.resources = {
@@ -302,6 +313,12 @@ class AmplifyFunction
     };
 
     this.storeOutput(outputStorageStrategy);
+
+    new AttributionMetadataStorage().storeAttributionMetadata(
+      Stack.of(this),
+      functionStackType,
+      fileURLToPath(new URL('../package.json', import.meta.url))
+    );
   }
 
   getResourceAccessAcceptor = () => ({