+++ packages/cli-core/src/package-manager-controller/yarn_classic_package_manager_controller.test.ts
@@ -148,7 +148,8 @@ void describe('YarnClassicPackageManagerController', () => {
           existsSyncMock
         );
       await yarnClassicPackageManagerController.initializeTsConfig('./amplify');
-      assert.equal(executeWithDebugLoggerMock.mock.callCount(), 2);
+      assert.equal(executeWithDebugLoggerMock.mock.callCount(), 1);
+      assert.equal(fspMock.writeFile.mock.callCount(), 1);
     });
   });
 });