+++ packages/cli-core/src/package-manager-controller/yarn_modern_package_manager_controller.test.ts
@@ -144,7 +144,7 @@ void describe('YarnModernPackageManagerController', () => {
         );
 
       await yarnModernPackageManagerController.initializeTsConfig('./amplify');
-      assert.equal(executeWithDebugLoggerMock.mock.callCount(), 2);
+      assert.equal(executeWithDebugLoggerMock.mock.callCount(), 1);
       assert.equal(fspMock.writeFile.mock.callCount(), 2);
     });
   });