+++ packages/cli/src/commands/sandbox/sandbox_event_handler_factory.ts
@@ -2,8 +2,7 @@ import { SandboxEventHandlerCreator } from './sandbox_command.js';
 import { BackendIdentifier } from '@aws-amplify/plugin-types';
 import { AmplifyError, UsageDataEmitter } from '@aws-amplify/platform-core';
 import { DeployResult } from '@aws-amplify/backend-deployer';
-import { COLOR } from '@aws-amplify/cli-core';
-import { printer } from '../../printer.js';
+import { format, printer } from '@aws-amplify/cli-core';
 
 /**
  * Coordinates creation of sandbox event handlers
@@ -46,18 +45,17 @@ export class SandboxEventHandlerFactory {
           } catch (error) {
             // Don't crash sandbox if config cannot be generated, but print the error message
             printer.print(
-              'Amplify configuration could not be generated.',
-              COLOR.RED
+              format.error('Amplify configuration could not be generated.')
             );
             if (error instanceof Error) {
-              printer.print(error.message, COLOR.RED);
+              printer.print(format.error(error.message));
             } else {
               try {
-                printer.print(JSON.stringify(error, null, 2), COLOR.RED);
+                printer.print(format.error(JSON.stringify(error, null, 2)));
               } catch {
                 // fallback in case there's an error stringify the error
                 // like with circular references.
-                printer.print('Unknown error', COLOR.RED);
+                printer.print(format.error('Unknown error'));
               }
             }
           }