+++ packages/create-amplify/src/gitignore_initializer.test.ts
@@ -3,7 +3,7 @@ import { GitIgnoreInitializer } from './gitignore_initializer.js';
 import assert from 'assert';
 import * as path from 'path';
 import * as os from 'os';
-import { printer } from './printer.js';
+import { printer } from '@aws-amplify/cli-core';
 
 void describe('GitIgnoreInitializer', () => {
   const logMock = mock.method(printer, 'log');