+++ packages/integration-tests/src/test-project-setup/data_storage_auth_with_triggers.ts
@@ -38,12 +38,8 @@ export class DataStorageAuthWithTriggerTestProjectCreator
   ) {}
 
   createProject = async (e2eProjectDir: string): Promise<TestProjectBase> => {
-    const {
-      projectName,
-      projectRoot,
-      projectAmplifyDir,
-      projectDotAmplifyDir,
-    } = await createEmptyAmplifyProject(this.name, e2eProjectDir);
+    const { projectName, projectRoot, projectAmplifyDir } =
+      await createEmptyAmplifyProject(this.name, e2eProjectDir);
 
     const project = new DataStorageAuthWithTriggerTestProject(
       projectName,
@@ -63,11 +59,6 @@ export class DataStorageAuthWithTriggerTestProjectCreator
         recursive: true,
       }
     );
-
-    // copy .amplify folder with typedef file from source project
-    await fs.cp(project.sourceProjectDotAmplifyDirPath, projectDotAmplifyDir, {
-      recursive: true,
-    });
 
     return project;
   };
@@ -97,11 +88,9 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
   );
 
   private readonly sourceProjectUpdateDirPath: URL = new URL(
-    `${this.sourceProjectDirPath}/update-1`,
+    `${this.sourceProjectDirPath}/hotswap-update-files`,
     import.meta.url
   );
-
-  private readonly dataResourceFileSuffix = 'data/resource.ts';
 
   private readonly testSecretNames = [
     'googleId',
@@ -165,19 +154,19 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
    * @inheritdoc
    */
   override async getUpdates(): Promise<TestProjectUpdate[]> {
-    const sourceDataResourceFile = pathToFileURL(
-      path.join(
-        fileURLToPath(this.sourceProjectUpdateDirPath),
-        this.dataResourceFileSuffix
-      )
-    );
-    const dataResourceFile = pathToFileURL(
-      path.join(this.projectAmplifyDirPath, this.dataResourceFileSuffix)
-    );
     return [
       {
-        sourceFile: sourceDataResourceFile,
-        projectFile: dataResourceFile,
+        replacements: [this.getUpdateReplacementDefinition('data/resource.ts')],
+        deployThresholdSec: {
+          onWindows: 40,
+          onOther: 30,
+        },
+      },
+      {
+        replacements: [
+          this.getUpdateReplacementDefinition('func-src/handler.ts'),
+          this.getUpdateReplacementDefinition('function.ts'),
+        ],
         deployThresholdSec: {
           onWindows: 40,
           onOther: 30,
@@ -244,6 +233,19 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
 
     assert.ok(typedShimStats.isFile());
   }
+
+  private getUpdateReplacementDefinition = (suffix: string) => ({
+    source: this.getSourcePath(suffix),
+    destination: this.getTestProjectPath(suffix),
+  });
+
+  private getSourcePath = (suffix: string) =>
+    pathToFileURL(
+      path.join(fileURLToPath(this.sourceProjectUpdateDirPath), suffix)
+    );
+
+  private getTestProjectPath = (suffix: string) =>
+    pathToFileURL(path.join(this.projectAmplifyDirPath, suffix));
 
   private setUpDeployEnvironment = async (
     backendId: BackendIdentifier