+++ packages/integration-tests/src/test-project-setup/setup_dir_as_esm_module.ts
@@ -1,6 +1,5 @@
 import * as path from 'path';
 import * as fs from 'fs/promises';
-import { execa } from 'execa';
 
 /**
  * Configures a minimal package.json and tsconfig.json to make the specified directory compile as an ESM module
@@ -12,37 +11,25 @@ export const setupDirAsEsmModule = async (absoluteDirPath: string) => {
     JSON.stringify(packageJsonContent, null, 2)
   );
 
-  const tscArgs = [
-    'tsc',
-    '--init',
-    '--resolveJsonModule',
-    'true',
-    '--module',
-    'node16',
-    '--moduleResolution',
-    'node16',
-    '--target',
-    'es2022',
-  ];
-
-  await execa('npx', tscArgs, {
-    stdio: 'inherit',
-    cwd: absoluteDirPath,
-  });
-
-  const pathsObj = {
-    // The path here is coupled with backend-function's generated typedef file path
-    '@env/*': ['../.amplify/function-env/*'],
+  const tsConfigTemplate = {
+    compilerOptions: {
+      target: 'es2022',
+      module: 'es2022',
+      moduleResolution: 'bundler',
+      resolveJsonModule: true,
+      // eslint-disable-next-line spellcheck/spell-checker
+      esModuleInterop: true,
+      forceConsistentCasingInFileNames: true,
+      strict: true,
+      skipLibCheck: true,
+      // The path here is coupled with backend-function's generated typedef file path
+      paths: { '@env/*': ['../.amplify/function-env/*'] },
+    },
   };
-
   const tsConfigPath = path.resolve(absoluteDirPath, 'tsconfig.json');
-  const tsConfigContent = (await fs.readFile(tsConfigPath, 'utf-8')).replace(
-    /\/\*[\s\S]*?\*\/|([^:]|^)\/\/.*$/gm, // Removes all comments
-    ''
+  await fs.writeFile(
+    tsConfigPath,
+    JSON.stringify(tsConfigTemplate, null, 2),
+    'utf-8'
   );
-  const tsConfigObject = JSON.parse(tsConfigContent);
-
-  // Add paths object and overwrite the tsconfig file
-  tsConfigObject.compilerOptions.paths = pathsObj;
-  await fs.writeFile(tsConfigPath, JSON.stringify(tsConfigObject), 'utf-8');
 };