+++ packages/integration-tests/src/test-projects/data-storage-auth-with-triggers-ts/update-1/function.ts
@@ -1,5 +1,0 @@
-import { defineFunction } from '@aws-amplify/backend';
-
-export const defaultNodeFunc = defineFunction({
-  entry: './func-src/handler.ts',
-});