+++ packages/auth-construct/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/auth-construct-alpha
 
+## 0.6.0-beta.5
+
+### Patch Changes
+
+- @aws-amplify/backend-output-storage@0.4.0-beta.2
+
+## 0.6.0-beta.4
+
+### Patch Changes
+
+- 1d444df: Fix deployment bug with SAML providers.
+
 ## 0.6.0-beta.3
 
 ### Patch Changes